+++ changelog.md
@@ -1,5 +1,9 @@
 ## Changelog
 
+-- 1.1.0 --
+* Changed how it shows states
+* Added Celestial Tournament
+
 -- 1.0.0 --
 * Created addon
 * added the 4 current pet dungeons.